serving-service/requirements.txt
python-dotenv
flask
autogluon
pandas
kagglehub
confluent_kafka
scikit-learn
flask-cors
bcrypt
flask-jwt-extended
cassandra-driver